+++ libs/components/src/lib/date-picker/date-picker.spec.ts
@@ -32,13 +32,20 @@ describe('vwc-date-picker', () => {
 
 	const getButtonByLabel = (label: string) =>
 		(element.shadowRoot!.querySelector(`[aria-label="${label}"]`) ??
-			element.shadowRoot!.querySelector(`[label="${label}"]`)) as Button;
+			element.shadowRoot!.querySelector(`[label="${label}"]`) ??
+			element.shadowRoot!.querySelector(
+				`[data-vvd-aria-label="${label}"]`
+			)) as Button;
 
 	const getDialogTitle = () => titleAction.textContent!.trim();
 
 	function typeIntoTextField(text: string) {
 		textField.value = text;
 		textField.dispatchEvent(new InputEvent('input'));
+	}
+
+	function typeIntoTextFieldAndBlur(text: string) {
+		typeIntoTextField(text);
 		textField.dispatchEvent(new InputEvent('change'));
 		textField.dispatchEvent(new Event('blur'));
 		element.dispatchEvent(new Event('focusout'));
@@ -96,7 +103,7 @@ describe('vwc-date-picker', () => {
 			(shadowRoot) => {
 				return {
 					firstFocusable: shadowRoot.querySelector(
-						'vwc-button[aria-label="Previous year"]'
+						'vwc-button[data-vvd-aria-label="Previous year"]'
 					)!,
 					lastFocusable: shadowRoot.querySelector('vwc-button[label="OK"]')!,
 				};
@@ -133,14 +140,14 @@ describe('vwc-date-picker', () => {
 		});
 
 		it('should update value when a user enters a valid date into the text field', async () => {
-			typeIntoTextField('01/21/2021');
+			typeIntoTextFieldAndBlur('01/21/2021');
 			await elementUpdated(element);
 
 			expect(element.value).toBe('2021-01-21');
 		});
 
 		it('should keep an empty value when a user enters a invalid date into the text field', async () => {
-			typeIntoTextField('x');
+			typeIntoTextFieldAndBlur('x');
 			await elementUpdated(element);
 
 			expect(element.value).toBe('');
@@ -149,7 +156,7 @@ describe('vwc-date-picker', () => {
 		it('should clear the value but keep invalid input when a user enters a invalid date into the text field', async () => {
 			element.value = '2021-01-21';
 
-			typeIntoTextField('x');
+			typeIntoTextFieldAndBlur('x');
 			await elementUpdated(element);
 
 			expect(element.value).toBe('');
@@ -165,6 +172,26 @@ describe('vwc-date-picker', () => {
 
 			expect(textField.value).toBe('');
 		});
+
+		describe('while typing into the text field', () => {
+			it('should update value if input is valid', async () => {
+				typeIntoTextField('1/21/2021');
+				await elementUpdated(element);
+
+				expect(element.value).toBe('2021-01-21');
+				expect(textField.value).toBe('1/21/2021');
+			});
+
+			it('should keep the previous value if input is invalid', async () => {
+				element.value = '2021-01-21';
+
+				typeIntoTextField('2');
+				await elementUpdated(element);
+
+				expect(element.value).toBe('2021-01-21');
+				expect(textField.value).toBe('2');
+			});
+		});
 	});
 
 	describe.each(['input', 'change'])('%s event', (eventName) => {
@@ -172,7 +199,7 @@ describe('vwc-date-picker', () => {
 			const spy = vi.fn();
 			element.addEventListener(eventName, spy);
 
-			typeIntoTextField('01/21/2021');
+			typeIntoTextFieldAndBlur('01/21/2021');
 			await elementUpdated(element);
 
 			expect(spy).toHaveBeenCalledTimes(1);
@@ -189,31 +216,53 @@ describe('vwc-date-picker', () => {
 		});
 	});
 
+	describe('input event', () => {
+		it('should be fired when the value updates while a user is typing', async () => {
+			const spy = vi.fn();
+			element.addEventListener('input', spy);
+
+			typeIntoTextField('01/21/2021');
+			await elementUpdated(element);
+
+			expect(spy).toHaveBeenCalledTimes(1);
+		});
+	});
+
+	describe('change event', () => {
+		it('should not be fired when the value updates while a user is typing', async () => {
+			const spy = vi.fn();
+			element.addEventListener('change', spy);
+
+			typeIntoTextField('01/21/2021');
+			await elementUpdated(element);
+
+			expect(spy).toHaveBeenCalledTimes(0);
+		});
+	});
+
 	describe('text field', () => {
 		it('should show an invalid date error when an invalid date is entered', async () => {
-			typeIntoTextField('invalid date');
-			textField.dispatchEvent(new Event('blur'));
+			typeIntoTextFieldAndBlur('invalid date');
 			await elementUpdated(element);
 
 			expect(textField.errorText).toBe('Please enter a valid date.');
 		});
 
 		it('should clear the invalid date error when a valid date is entered', async () => {
-			typeIntoTextField('invalid date');
-			textField.dispatchEvent(new Event('blur'));
+			typeIntoTextFieldAndBlur('invalid date');
 			await elementUpdated(element);
 
-			typeIntoTextField('01/21/2021');
+			typeIntoTextFieldAndBlur('01/21/2021');
 			await elementUpdated(element);
 
 			expect(textField.errorText).toBe('');
 		});
 
 		it('should clear the value when an empty string is entered', async () => {
-			typeIntoTextField('01/21/2021');
+			typeIntoTextFieldAndBlur('01/21/2021');
 			await elementUpdated(element);
 
-			typeIntoTextField('');
+			typeIntoTextFieldAndBlur('');
 			await elementUpdated(element);
 
 			expect(textField.value).toBe('');
@@ -251,16 +300,14 @@ describe('vwc-date-picker', () => {
 		});
 
 		it('should have an aria-label of "Choose date" when no date is selected', async () => {
-			expect(pickerButton.getAttribute('aria-label')).toBe('Choose date');
+			expect(pickerButton.ariaLabel).toBe('Choose date');
 		});
 
 		it('should have an aria-label of "Change date, DATE" when a date is selected', async () => {
 			element.value = '2021-01-01';
 			await elementUpdated(element);
 
-			expect(pickerButton.getAttribute('aria-label')).toBe(
-				'Change date, 01/01/2021'
-			);
+			expect(pickerButton.ariaLabel).toBe('Change date, 01/01/2021');
 		});
 	});
 
@@ -270,7 +317,7 @@ describe('vwc-date-picker', () => {
 		});
 
 		it('should update current month to the month of the selected date when the user enters a new date', async () => {
-			typeIntoTextField('01/21/2021');
+			typeIntoTextFieldAndBlur('01/21/2021');
 			await elementUpdated(element);
 
 			expect(getDialogTitle()).toBe('January 2021');